+++ packages/components/RichTextEditor/index.ts
@@ -1,0 +1,4 @@
+export { RichTextEditor } from './RichTextEditor'
+export type { RichTextEditorProps } from './RichTextEditor'
+export { RichTextEditorDemo } from './RichTextEditorDemo'
+export { defaultToolbar } from './toolbar'