+++ src/pages/index/demo/components/rich-text-editor.tsx
@@ -1,0 +1,3 @@
+import { ZaRichTextEditorDemo } from '@zealous-admin/components/index'
+
+export default ZaRichTextEditorDemo